fixed join vs aggregates; fixed filters on aliased joined attrs
