+++ src/sphinxint.h
@@ -144,78 +144,6 @@ enum QueryDebug_e
 	QUERY_DEBUG_NO_PAYLOAD = 1<<0
 };
 
-class Docstore_i;
-
-/// per-query search context
-/// everything that index needs to compute/create to process the query
-class CSphQueryContext : public ISphNoncopyable
-{
-public:
-	// searching-only, per-query
-	const CSphQuery &			m_tQuery;
-
-	int							m_iWeights = 0;					///< search query field weights count
-	int							m_dWeights [ SPH_MAX_FIELDS ];	///< search query field weights
-
-	DWORD						m_uPackedFactorFlags { SPH_FACTOR_DISABLE }; ///< whether we need to calculate packed factors (and some extra options)
-
-	std::unique_ptr<ISphFilter>	m_pFilter;
-	std::unique_ptr<ISphFilter>	m_pWeightFilter;
-
-	bool						m_bSkipQCache = false;			///< whether do not cache this query
-
-	struct CalcItem_t
-	{
-		CSphAttrLocator			m_tLoc;					///< result locator
-		ESphAttr				m_eType { SPH_ATTR_NONE};	///< result type
-		ISphExprRefPtr_c		m_pExpr;		///< evaluator (non-owned)
-	};
-	CSphVector<CalcItem_t>		m_dCalcFilter;			///< items to compute for filtering
-	CSphVector<CalcItem_t>		m_dCalcSort;			///< items to compute for sorting/grouping
-	CSphVector<CalcItem_t>		m_dCalcFinal;			///< items to compute when finalizing result set
-
-	IntVec_t					m_dCalcFilterPtrAttrs;	///< items to free after computing filter stage
-	IntVec_t					m_dCalcSortPtrAttrs;	///< items to free after computing sort stage
-
-	const void *							m_pIndexData = nullptr;	///< backend specific data
-	QueryProfile_c *						m_pProfile = nullptr;
-	const SmallStringHash_T<int64_t> *		m_pLocalDocs = nullptr;
-	int64_t									m_iTotalDocs = 0;
-	int64_t									m_iIndexTotalDocs = 0;
-
-public:
-	explicit CSphQueryContext ( const CSphQuery & q );
-			~CSphQueryContext ();
-
-	void	BindWeights ( const CSphQuery & tQuery, const CSphSchema & tSchema, CSphString & sWarning );
-	bool	SetupCalc ( CSphQueryResultMeta & tMeta, const ISphSchema & tInSchema, const ISphSchema & tSchema, const BYTE * pBlobPool, const columnar::Columnar_i * pColumnar, const CSphVector<const ISphSchema *> & dInSchemas );
-	bool	CreateFilters ( CreateFilterContext_t& tCtx, CSphString &sError, CSphString &sWarning );
-
-	void	CalcFilter ( CSphMatch & tMatch ) const;
-	void	CalcSort ( CSphMatch & tMatch ) const;
-	void	CalcFinal ( CSphMatch & tMatch ) const;
-	void	CalcItem ( CSphMatch & tMatch, const CalcItem_t & tCalc ) const;
-
-	void	FreeDataFilter ( CSphMatch & tMatch ) const;
-	void	FreeDataSort ( CSphMatch & tMatch ) const;
-
-	// note that RT index bind pools at segment searching, not at time it setups context
-	void	ExprCommand ( ESphExprCommand eCmd, void * pArg );
-	void	SetBlobPool ( const BYTE * pBlobPool );
-	void	SetColumnar ( const columnar::Columnar_i * pColumnar );
-	void	SetDocstore ( const Docstore_i * pDocstore, int64_t iDocstoreSessionId );
-
-	void	SetupExtraData ( ISphRanker * pRanker, ISphMatchSorter * pSorter );
-	void	ResetFilters();
-
-private:
-	CSphVector<UservarIntSet_c>		m_dUserVals;
-
-	void	AddToFilterCalc ( const CalcItem_t & tCalc );
-	void	AddToSortCalc ( const CalcItem_t & tCalc );
-};
-
-
 // collect valid schemas from sorters, excluding one
 CSphVector<const ISphSchema *> SorterSchemas ( const VecTraits_T<ISphMatchSorter *> & dSorters, int iSkipSorter );
