+++ src/stackmock.cpp
@@ -16,6 +16,7 @@
 #include "coroutine.h"
 #include "searchdsql.h"
 #include "attribute.h"
+#include "querycontext.h"
 
 // hard-coded definitions to avoid probing (that is - to avoid confusing memcheck programs)
 // run searchd with --logdebug --console once, read values, then write them here and uncomment these lines
